refactor: Be backend agnostic
It strikes me that I'm getting pretty fond of the user interface for
bugger, and would likely want to reuse it in the future for other
trackers.

Thus, this patch adds some level of opaqueness in terms of which backend
is used. It's still not configurable which backend to use, but that
should be easy enough to add when I need it.

Mechanized Mantis was a cool alliteration, though. *sob*